Split analyzer utils
